Use .vue.html extension for AoT-ing Vue.js HTML templates!
